requirements.txt
streamlit==1.22.0
plotly==5.9.0
pandas==1.5.3
numpy==1.23.5
folium==0.14.0
haversine==2.8.0
streamlit-folium==0.11.1
Pillow==9.4.0